[3.0] remove php template helpers
