+++ src/CoreShop/Bundle/AddressBundle/Resources/config/services.yml
@@ -46,18 +46,10 @@ services:
         arguments:
             - '@pimcore.translator'
 
-    CoreShop\Bundle\AddressBundle\Templating\Helper\FormatAddressHelperInterface: '@CoreShop\Bundle\AddressBundle\Templating\Helper\FormatAddressHelper'
-    CoreShop\Bundle\AddressBundle\Templating\Helper\FormatAddressHelper:
-        arguments:
-            - '@CoreShop\Component\Address\Formatter\AddressFormatter'
-        lazy: true
-        tags:
-            - { name: templating.helper, alias: coreshop_format_address }
-
     CoreShop\Bundle\AddressBundle\Twig\FormatAddressExtension:
         public: false
         arguments:
-            - '@CoreShop\Bundle\AddressBundle\Templating\Helper\FormatAddressHelper'
+            - '@CoreShop\Component\Address\Formatter\AddressFormatterInterface'
         tags:
             - { name: twig.extension }
