--- src/CoreShop/Bundle/CoreBundle/Templating/Helper/ValidPriceRulesHelper.php
@@ -1,46 +1,0 @@
-<?php
-/**
- * CoreShop.
- *
- * This source file is subject to the GNU General Public License version 3 (GPLv3)
- * For the full copyright and license information, please view the LICENSE.md and gpl-3.0.txt
- * files that are distributed with this source code.
- *
- * @copyright  Copyright (c) 2015-2020 Dominik Pfaffenbauer (https://www.pfaffenbauer.at)
- * @license    https://www.coreshop.org/license     GNU General Public License version 3 (GPLv3)
- */
-
-declare(strict_types=1);
-
-namespace CoreShop\Bundle\CoreBundle\Templating\Helper;
-
-use CoreShop\Component\Core\Context\ShopperContextInterface;
-use CoreShop\Component\Product\Model\ProductInterface;
-use CoreShop\Component\Product\Rule\Fetcher\ValidRulesFetcherInterface;
-use Symfony\Component\Templating\Helper\Helper;
-
-class ValidPriceRulesHelper extends Helper implements ValidPriceRulesHelperInterface
-{
-    protected $validPriceRulesFetcher;
-
-    public function __construct(ValidRulesFetcherInterface $validPriceRulesFetcher)
-    {
-        $this->validPriceRulesFetcher = $validPriceRulesFetcher;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getValidRules(ProductInterface $product, array $context = []): array
-    {
-        return $this->validPriceRulesFetcher->getValidRules($product, $context);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getName()
-    {
-        return 'coreshop_product_price_rules';
-    }
-}
