+++ src/CoreShop/Bundle/CoreBundle/Twig/ProductQuantityPriceRuleRangesPriceExtension.php
@@ -15,16 +15,39 @@ declare(strict_types=1);
 namespace CoreShop\Bundle\CoreBundle\Twig;
 
 use CoreShop\Bundle\CoreBundle\Templating\Helper\ProductQuantityPriceRuleRangesPriceHelperInterface;
+use CoreShop\Component\Address\Model\AddressInterface;
+use CoreShop\Component\Core\Model\ProductInterface;
+use CoreShop\Component\Core\Model\QuantityRangeInterface;
+use CoreShop\Component\Core\Product\ProductTaxCalculatorFactoryInterface;
+use CoreShop\Component\Core\Provider\DefaultTaxAddressProviderInterface;
+use CoreShop\Component\Core\Taxation\TaxApplicatorInterface;
+use CoreShop\Component\Order\Calculator\PurchasableCalculatorInterface;
+use CoreShop\Component\Order\Model\PurchasableInterface;
+use CoreShop\Component\ProductQuantityPriceRules\Detector\QuantityReferenceDetectorInterface;
+use CoreShop\Component\Taxation\Calculator\TaxCalculatorInterface;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFunction;
 
 final class ProductQuantityPriceRuleRangesPriceExtension extends AbstractExtension
 {
-    private $helper;
+    protected $quantityReferenceDetector;
+    protected $purchasableCalculator;
+    private $defaultTaxAddressProvider;
+    private $taxCalculatorFactory;
+    private $taxApplicator;
 
-    public function __construct(ProductQuantityPriceRuleRangesPriceHelperInterface $helper)
-    {
-        $this->helper = $helper;
+    public function __construct(
+        QuantityReferenceDetectorInterface $quantityReferenceDetector,
+        PurchasableCalculatorInterface $purchasableCalculator,
+        DefaultTaxAddressProviderInterface $defaultTaxAddressProvider,
+        ProductTaxCalculatorFactoryInterface $taxCalculatorFactory,
+        TaxApplicatorInterface $taxApplicator
+    ) {
+        $this->quantityReferenceDetector = $quantityReferenceDetector;
+        $this->purchasableCalculator = $purchasableCalculator;
+        $this->defaultTaxAddressProvider = $defaultTaxAddressProvider;
+        $this->taxCalculatorFactory = $taxCalculatorFactory;
+        $this->taxApplicator = $taxApplicator;
     }
 
     /**
@@ -33,7 +56,35 @@ final class ProductQuantityPriceRuleRangesPriceExtension extends AbstractExtensi
     public function getFunctions(): array
     {
         return [
-            new TwigFunction('coreshop_quantity_price_rule_range_price', [$this->helper, 'getQuantityPriceRuleRangePrice']),
+            new TwigFunction('coreshop_quantity_price_rule_range_price', [$this, 'getQuantityPriceRuleRangePrice']),
         ];
+    }
+
+    public function getQuantityPriceRuleRangePrice(
+        QuantityRangeInterface $range,
+        ProductInterface $product,
+        array $context,
+        bool $withTax = true
+    ): int {
+        $realItemPrice = $this->purchasableCalculator->getPrice($product, $context);
+        $price = $this->quantityReferenceDetector->detectRangePrice($product, $range, $realItemPrice, $context);
+
+        $taxCalculator = $this->getTaxCalculator($product, $context);
+
+        if ($taxCalculator instanceof TaxCalculatorInterface) {
+            return $this->taxApplicator->applyTax($price, $context, $taxCalculator, $withTax);
+        }
+
+        return $price;
+    }
+
+    protected function getTaxCalculator(PurchasableInterface $product, array $context): ?TaxCalculatorInterface
+    {
+        return $this->taxCalculatorFactory->getTaxCalculator($product, $this->getDefaultAddress($context));
+    }
+
+    protected function getDefaultAddress(array $context): ?AddressInterface
+    {
+        return $this->defaultTaxAddressProvider->getAddress($context);
     }
 }
