+++ src/CoreShop/Bundle/CoreBundle/Twig/ProductTaxExtension.php
@@ -15,16 +15,29 @@ declare(strict_types=1);
 namespace CoreShop\Bundle\CoreBundle\Twig;
 
 use CoreShop\Bundle\CoreBundle\Templating\Helper\ProductTaxHelperInterface;
+use CoreShop\Component\Core\Product\ProductTaxCalculatorFactoryInterface;
+use CoreShop\Component\Core\Product\TaxedProductPriceCalculatorInterface;
+use CoreShop\Component\Core\Provider\DefaultTaxAddressProviderInterface;
+use CoreShop\Component\Order\Model\PurchasableInterface;
+use CoreShop\Component\Taxation\Calculator\TaxCalculatorInterface;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFilter;
 
 final class ProductTaxExtension extends AbstractExtension
 {
-    private $productTaxHelper;
+    private $productPriceCalculator;
+    private $taxCalculatorFactory;
+    private $defaultAddressProvider;
 
-    public function __construct(ProductTaxHelperInterface $productTaxHelper)
+    public function __construct(
+        TaxedProductPriceCalculatorInterface $productPriceCalculator,
+        ProductTaxCalculatorFactoryInterface $taxCalculatorFactory,
+        DefaultTaxAddressProviderInterface $defaultAddressProvider
+    )
     {
-        $this->productTaxHelper = $productTaxHelper;
+        $this->productPriceCalculator = $productPriceCalculator;
+        $this->taxCalculatorFactory = $taxCalculatorFactory;
+        $this->defaultAddressProvider = $defaultAddressProvider;
     }
 
     /**
@@ -33,8 +46,30 @@ final class ProductTaxExtension extends AbstractExtension
     public function getFilters(): array
     {
         return [
-            new TwigFilter('coreshop_product_tax_amount', [$this->productTaxHelper, 'getTaxAmount']),
-            new TwigFilter('coreshop_product_tax_rate', [$this->productTaxHelper, 'getTaxRate']),
+            new TwigFilter('coreshop_product_tax_amount', [$this, 'getTaxAmount']),
+            new TwigFilter('coreshop_product_tax_rate', [$this, 'getTaxRate']),
         ];
+    }
+
+    public function getTaxAmount(PurchasableInterface $product, array $context = []): int
+    {
+        $taxCalculator = $this->taxCalculatorFactory->getTaxCalculator($product, $this->defaultAddressProvider->getAddress($context));
+
+        if ($taxCalculator instanceof TaxCalculatorInterface) {
+            return $taxCalculator->getTaxesAmount($this->productPriceCalculator->getPrice($product, $context, false));
+        }
+
+        return 0;
+    }
+
+    public function getTaxRate(PurchasableInterface $product, array $context = []): float
+    {
+        $taxCalculator = $this->taxCalculatorFactory->getTaxCalculator($product, $this->defaultAddressProvider->getAddress($context));
+
+        if ($taxCalculator instanceof TaxCalculatorInterface) {
+            return $taxCalculator->getTotalRate();
+        }
+
+        return 0;
     }
 }
