+++ src/CoreShop/Bundle/CoreBundle/Twig/ValidPriceRulesExtension.php
@@ -15,16 +15,17 @@ declare(strict_types=1);
 namespace CoreShop\Bundle\CoreBundle\Twig;
 
 use CoreShop\Bundle\CoreBundle\Templating\Helper\ValidPriceRulesHelperInterface;
+use CoreShop\Component\Product\Rule\Fetcher\ValidRulesFetcherInterface;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFilter;
 
 final class ValidPriceRulesExtension extends AbstractExtension
 {
-    private $helper;
+    protected $validPriceRulesFetcher;
 
-    public function __construct(ValidPriceRulesHelperInterface $helper)
+    public function __construct(ValidRulesFetcherInterface $validPriceRulesFetcher)
     {
-        $this->helper = $helper;
+        $this->validPriceRulesFetcher = $validPriceRulesFetcher;
     }
 
     /**
@@ -33,7 +34,7 @@ final class ValidPriceRulesExtension extends AbstractExtension
     public function getFilters(): array
     {
         return [
-            new TwigFilter('coreshop_product_price_rules', [$this->helper, 'getValidRules']),
+            new TwigFilter('coreshop_product_price_rules', [$this->validPriceRulesFetcher, 'getValidRules']),
         ];
     }
 }
