+++ src/CoreShop/Bundle/CurrencyBundle/Twig/ConvertCurrencyExtension.php
@@ -15,16 +15,17 @@ declare(strict_types=1);
 namespace CoreShop\Bundle\CurrencyBundle\Twig;
 
 use CoreShop\Bundle\CurrencyBundle\Templating\Helper\ConvertCurrencyHelperInterface;
+use CoreShop\Component\Currency\Converter\CurrencyConverterInterface;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFilter;
 
 final class ConvertCurrencyExtension extends AbstractExtension
 {
-    private $helper;
+    private $currencyConverter;
 
-    public function __construct(ConvertCurrencyHelperInterface $helper)
+    public function __construct(CurrencyConverterInterface $currencyConverter)
     {
-        $this->helper = $helper;
+        $this->currencyConverter = $currencyConverter;
     }
 
     /**
@@ -33,7 +34,16 @@ final class ConvertCurrencyExtension extends AbstractExtension
     public function getFilters()
     {
         return [
-            new TwigFilter('coreshop_convert_currency', [$this->helper, 'convertAmount']),
+            new TwigFilter('coreshop_convert_currency', [$this, 'convertAmount']),
         ];
+    }
+
+    public function convertAmount(?int $amount, string $sourceCurrencyCode, string $targetCurrencyCode): int
+    {
+        if (null === $amount) {
+            return 0;
+        }
+
+        return $this->currencyConverter->convert($amount, $sourceCurrencyCode, $targetCurrencyCode);
     }
 }
