+++ src/CoreShop/Bundle/CurrencyBundle/Twig/CurrencyExtension.php
@@ -15,25 +15,25 @@ declare(strict_types=1);
 namespace CoreShop\Bundle\CurrencyBundle\Twig;
 
 use CoreShop\Bundle\CurrencyBundle\Templating\Helper\CurrencyHelperInterface;
+use Symfony\Component\Intl\Currencies;
+use Symfony\Component\Intl\Intl;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFilter;
 
 final class CurrencyExtension extends AbstractExtension
 {
-    private $helper;
-
-    public function __construct(CurrencyHelperInterface $helper)
-    {
-        $this->helper = $helper;
-    }
-
     /**
      * {@inheritdoc}
      */
     public function getFilters()
     {
         return [
-            new TwigFilter('coreshop_currency_symbol', [$this->helper, 'convertCurrencyCodeToSymbol']),
+            new TwigFilter('coreshop_currency_symbol', [$this, 'convertCurrencyCodeToSymbol']),
         ];
+    }
+
+    public function convertCurrencyCodeToSymbol(string $code, ?string $locale = null): string
+    {
+        return Currencies::getSymbol($code);
     }
 }
