+++ src/CoreShop/Bundle/InventoryBundle/Resources/config/services.yml
@@ -14,16 +14,9 @@ services:
         tags:
            - { name: validator.constraint_validator, alias: coreshop_in_stock }
 
-    CoreShop\Bundle\InventoryBundle\Templating\Helper\InventoryHelperInterface: '@CoreShop\Bundle\InventoryBundle\Templating\Helper\InventoryHelper'
-    CoreShop\Bundle\InventoryBundle\Templating\Helper\InventoryHelper:
-        arguments:
-          - '@CoreShop\Component\Inventory\Checker\AvailabilityCheckerInterface'
-        tags:
-            - { name: templating.helper, alias: coreshop_inventory }
-
     CoreShop\Bundle\InventoryBundle\Twig\InventoryExtension:
         public: false
         arguments:
-          - '@CoreShop\Bundle\InventoryBundle\Templating\Helper\InventoryHelper'
+          - '@CoreShop\Component\Inventory\Checker\AvailabilityCheckerInterface'
         tags:
             - { name: twig.extension }
