+++ src/CoreShop/Bundle/InventoryBundle/Twig/InventoryExtension.php
@@ -15,16 +15,17 @@ declare(strict_types=1);
 namespace CoreShop\Bundle\InventoryBundle\Twig;
 
 use CoreShop\Bundle\InventoryBundle\Templating\Helper\InventoryHelper;
+use CoreShop\Component\Inventory\Checker\AvailabilityCheckerInterface;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFunction;
 
 final class InventoryExtension extends AbstractExtension
 {
-    private $helper;
+    private $checker;
 
-    public function __construct(InventoryHelper $helper)
+    public function __construct(AvailabilityCheckerInterface $checker)
     {
-        $this->helper = $helper;
+        $this->checker = $checker;
     }
 
     /**
@@ -33,8 +34,8 @@ final class InventoryExtension extends AbstractExtension
     public function getFunctions()
     {
         return [
-            new TwigFunction('coreshop_inventory_is_available', [$this->helper, 'isStockAvailable']),
-            new TwigFunction('coreshop_inventory_is_sufficient', [$this->helper, 'isStockSufficient']),
+            new TwigFunction('coreshop_inventory_is_available', [$this->checker, 'isStockAvailable']),
+            new TwigFunction('coreshop_inventory_is_sufficient', [$this->checker, 'isStockSufficient']),
         ];
     }
 }
